obj_change/requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.20.0
tqdm>=4.60.0
faiss-gpu>=1.7.0  # GPUが使用できない場合は faiss-cpu に変更
Pillow>=8.0.0
ultralytics>=8.0.0  # FastSAMのために必要
timm>=0.6.0  # Depth Anything V2のために必要 